Changed how tokens and tags are put together
Changed token to accept an int value for the tag. The tag
is then referenced by the int that was entered. This was done
to allow character values such as '(' and ')' to correspond
to a Tag and a character.

Tags Enum was also changed. Most values start at 256 and move up,
however all single character Tags were renumbered based on their ascii
value

Added a Lexeme value to a token. This will be a single character for tokens
but for Words it will be a string of all of the letters that make up the
lexeme