Added Numerical Scanner FSA / Tried to fix other
We got the logic figured out for the Numerical (int/float/fixed) values. HOWEVER, once we went to test it, we found there was a deeper rooted problem. It appears the current character doesn't like numbers for some reason and skips over them. This could be a problem having to do with ascii numbers representing different things, but that seems odd. It may also be linked to the problem with the columns and line numbers not calculating properly as well.