changed IdentifierRecord to have a VariableSymbol
There was a rather sloppy implementation of getting the offset of a symbol which forced a offset property in the base symbol class when it was not going to be used in all of the inherited classes. Changed the IdentifierRecord to return a variable symbol and took the offset out of the symbol class. Also used this implementation to pass up the result factor record from the GenerateIdPush method in the semantic analyzer